started implementing ball identification

+++ client/robot_code/source/utils.c
@@ -11,12 +11,12 @@
 
 /// @brief Structure used to store the data collected from the sonar sensor
 struct Ray {
-    float distance;
+    int distance;
     time_t time;
     int angle;
 };
 
-void initRay(struct Ray* ray, float distance, int angle) {
+void initRay(struct Ray* ray, int distance, int angle) {
     ray->distance = distance;
     ray->time = time(NULL); //TODO: check if this gives time is ms, otherwise fix it
     ray->angle = angle;
